feat: 사이트 관리자 페이지 구현
1. 대시보드 정보 조회 및 그래프 표현 구현
2. 쿠폰 등록 기능 및 쿠폰 리스트 정보 출력 구현
3. 사이트 이용자(유저, 판매자) 조회 페이지 구현
4. 판매자 등록 기능 구현
5. 사이트 관리자 등록 기능 구현

+++ src/main/java/kr/co/swm/model/service/SellerService.java
@@ -9,52 +9,52 @@ public interface SellerService {
 //  □□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□
 
     // 업소 조회수
-    int roomViews(int accommodationNo);
+    int roomViews(Long accommodationNo);
 
     // 대시보드 정보 조회
-    List<SellerDto> mainList(int accommodationNo);
+    List<SellerDto> mainList(Long accommodationNo);
 
 //  □□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□
 
     // 예약 검색 조회
-    List<SellerDto> reservationSearch(int accommodationNo, String dateType, String startDate, String endDate, String searchKeyword, String reservationStatus);
+    List<SellerDto> reservationSearch(Long accommodationNo, String dateType, String startDate, String endDate, String searchKeyword, String reservationStatus);
 
     // 일별 예약 조회
-    List<SellerDto> roomData(int accommodationNo, String selectedDate);
+    List<SellerDto> roomData(Long accommodationNo, String selectedDate);
 
     // 관리자 보유 객실 조회
-    List<SellerDto> accommodationRoomData(int accommodationNo);
+    List<SellerDto> accommodationRoomData(Long accommodationNo);
 
     // 월별 예약 조회
-    List<SellerDto> monthlyData(int accommodationNo);
+    List<SellerDto> monthlyData(Long accommodationNo);
 
 //  □□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□
 
     // 객실 리스트 조회
-    List<String> roomNameSearch(int accommodationNo);
+    List<String> roomNameSearch(Long accommodationNo);
 
     // 기본 요금 조회
-    List<SellerDto> basicRateList(String roomName, int accommodationNo);
+    List<SellerDto> basicRateList(String roomName, Long accommodationNo);
 
     // 추가 요금 조회
-    List<SellerDto> extraRateList(String roomName, int accommodationNo);
+    List<SellerDto> extraRateList(String roomName, Long accommodationNo);
 
     // 기본 요금 업데이트
-    int basicRateUpdate(SellerDto sellerDto, int accommodationNo);
+    int basicRateUpdate(SellerDto sellerDto, Long accommodationNo);
 
     // 추가 요금 업데이트 또는 삽입
-    int extraRateUpdate(SellerDto sellerDto, int accommodationNo);
+    int extraRateUpdate(SellerDto sellerDto, Long accommodationNo);
 
 
 //  □□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□□
 
     // 추가 요금 관련 리스트 조회
-    List<SellerDto.ExtraDto> extraSeasonList(int accommodationNo);
+    List<SellerDto.ExtraDto> extraSeasonList(Long accommodationNo);
 
     // 추가 요금 항목 삭제(extraName 기준)
-    boolean extraRateDelete(String extraName, int accommodationNo);
+    boolean extraRateDelete(String extraName, Long accommodationNo);
 
     // 추가 요금 기간 수정(extraDateStart, extraDateEnd)
-    boolean extraSeasonUpdate(List<SellerDto.ExtraDto> extraSeasonUpdate, int accommodationNo);
+    boolean extraSeasonUpdate(List<SellerDto.ExtraDto> extraSeasonUpdate, Long accommodationNo);
 
 }
